add super class for OOP
