+++ src/main/java/Character/Chr.java
@@ -1,0 +1,78 @@
+package Character;
+
+import Main.View;
+import Util.GameObject;
+import Util.Time;
+
+import java.awt.*;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
+
+public class Chr extends GameObject implements KeyListener {
+    protected int life = 3;
+
+    private double jumpingDelay = 0.020;
+    private double landingDelay = 0.020;
+
+    private boolean jumping = false;
+    private boolean landing = false;
+
+    private int jumpLimit = 150;
+    protected int ground_y = 500;
+
+    public Chr(View view) {
+        super(view);
+        gap = 20;
+    }
+
+    @Override
+    public void draw (Graphics g, View view) {
+        super.draw(g, view);
+        Jump();
+    }
+
+    @Override
+    public void keyTyped(KeyEvent e) {
+
+    }
+
+    @Override
+    public void keyPressed(KeyEvent e) {
+        switch(e.getKeyCode())
+        {
+            case KeyEvent.VK_UP:
+                if (!jumping && !landing)
+                    jumping = true;
+            case KeyEvent.VK_SPACE:
+                if (!jumping && !landing)
+                    jumping = true;
+                break;
+        }
+        //System.out.println(x+", "+y);
+    }
+
+    @Override
+    public void keyReleased(KeyEvent e) {
+
+    }
+
+     protected void Jump() {
+        if (jumping) {
+            if (time.timeCtrl(jumpingDelay)) // 점프
+                pos_y -= gap;
+            if (pos_y < ground_y - jumpLimit) {
+                jumping = false;
+                landing = true;
+            }
+        }
+        else if (landing) {
+            if (time.timeCtrl(landingDelay)) {      // 착지
+                pos_y += gap;
+            }
+            if (pos_y == ground_y) {
+                jumping = false;
+                landing = false;
+            }
+        }
+    }
+}
