+++ src/main/java/Item/Item.java
@@ -1,0 +1,10 @@
+package Item;
+
+import Main.View;
+import Util.GameObject;
+
+public class Item extends GameObject {
+    public Item (View view) {
+        super(view);
+    }
+}
